moving shallow_water.hpp to libmpdata++/solver so that library users using shallow_water equations do not have to ship their own .hpp file
